[PAGOPA-954] feat: added K6 performance tests
