+++ performance-test/src/environments/local.environment.json
@@ -1,0 +1,11 @@
+{
+  "environment": [
+    {
+      "env": "local",
+      "app_host": "http://localhost:8080",
+      "id_broker": "11111111111",
+      "id_station": "11111111111_01",
+      "id_ci": "09633951000"
+    }
+  ]
+}
