+++ performance-test/src/environments/uat.environment.json
@@ -1,0 +1,11 @@
+{
+  "environment": [
+    {
+      "env": "uat",
+      "app_host": "https://api.uat.platform.pagopa.it/apiconfig-selfcare-integration/v1",
+      "id_broker": "11111111111",
+      "id_station": "11111111111_01",
+      "id_ci": "09633951000"
+    }
+  ]
+}
