--- performance-test/src/local.environment.json
@@ -1,9 +1,0 @@
-{
-  "environment": [
-    {
-      "env": "local",
-      "host": "http://localhost:8080",
-      "basePath": ""
-    }
-  ]
-}
